refactor(core): package project tour orchestration
