+++ framework/core/src/python/kungfu/project_tour/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: Apache-2.0
+
+"""Project Tour runtime package."""
